Refactor references after choosing a name for the bridge
