+++ examples/partialorderplan.py
@@ -33,8 +33,8 @@ from parallel import (
 )
 from unified_planning.shortcuts import PlanKind
 
-from up_bridge.bridge import Bridge
-from up_bridge.executor import Executor
+from up_esb.bridge import Bridge
+from up_esb.executor import Executor
 
 # Define the problem. The domain is re-used from examples/parallel.py
